pyproject.toml
[project]
name = "agent-orchestration"
version = "0.1.0"
description = "Tinkering with Agent Orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pgvector>=0.4.1",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.35.0",
    "pydantic-settings",
    "python-json-logger",
]
